Support fetching queries to be verified using a Presto query
Currently, we support fetch source queries (pairs for queries to be
verified) from an MySQL table by a given suite. This is insufficient
in some cases, as it requires MySQL and is also bound by the limitation
of MySQL.

Add support to fetch source queries by running a Presto query. The
fetching query will be run on the helper cluster.