Defer the creation of dictionary in SliceDictionarySelectiveReader
Previously the dicitonary data byte arrays were allocated as soon as a
stripe or a rowgroup is open during read(). This commit persists them
as local buffers, and only copies them to create the dictionary until
getBlock() is called. This way the expensive allocations can be avoided
if a lazyBlock doesn't need to be read.

Moreover, this commit avoids creating Slices when evaluating the
filters, which was also another source of big  memory allocations.