Introduce large dictionary mode in SliceDictionarySelectiveReader
Previously we always allocate a dictionary for every rowgroup. When
these dictionaries are humongous, the allocations could cause reliability
and performance issues. This commit materializes the dictionaries if
they are too large. Instead of outputting a DictionaryBlock, it will
output a plain VariableWidthBlock if the dictionaries size is above
certain threshold. The experiment on user reported query shows over 10x
reduction in allocations and over 2x CPU reduction in scan.