Optimize MetadataQueryOptimizer with filter pushdown enabled
When hive.pushdown_filter_enabled is true, TableScanNode would
not be aware of table layout until connector plan optimizer finished
one pass. This would lead to inefficiency in MetadataQueryOptimizer
as it would try to get all the partitions of the table.

To optimize it, we moved MetadataQueryOptimizer after first connector
query optimizer run, and fixed the optimizer so it could adapt to
the layout after filter pushdown.