Fix local exchanges when materializing remote exchanges
Remote streaming exchanges enforce a "fixed" distribution, but table scans
only enforce a "multiple" distribution.  This meant we were missing local
exchanges in some cases when we materialized exchanges but required a fixed
distribution.  In particular this fixes join spilling when using
materialized exchanges, since in that case the probe side requires a
fixed distribution.